Further development of the effective mass calculation.
The effective mass should use MC truth only, the resolution smearing is taken into
account by using the reco information when filling the 'detected' events at a later stage.

In the future, we may want to have events shower_reco_good OR track_reco_good, for that purpose
there gandalf/shower hists persist in the output.

Added macro EffMass that takes the output of EffMhists and calculates the effective mass.

Added an option to run with a custom generation volume, useful for studies, mostly.