Rewrite and update all NO/IO asimov scripts
I have rewritten a whole part of the naming convention and the way
things are handled in the NO and IO scripts. This is the first
step towards unifying all of the code into one script that takes
input and does the type of calculation the user wants. Rewriting
everything takes too much time at this moment, ideally this is to
be done after the collabo meeting.