Work towards virtualisation of a small selection of functions.
Changes in the passing of parameters between FitPDF and FitUtil class. The aim is that `RecoEvts` and `TrueEvts` can be virtualised and classes can inherit from FitUtil, implement a new parameter and override e.g. `TrueEvts`, such that no modifications are required to `FitPDF` class and the base class `FitUtil`.