Add rangeName fnality to FitUtil::Expectation
The Expectation value function of FitUtil can now take a rangeName
string and it will go throught he E/ct/by variables and call the
associated range if it is defined. This range is then applied to the
generation of the expectation value histogram. In this way one can
bound the histogram when making it, and simultaneously this function
is also used by the fitter as integrator.

So now when one defines ranges, these can be passed to the
`RooSimultaneous->fitTo` method using the optional argument
Range("rangeName") and the pdfs will be simultaneously fitted on the
defined ranges "rangeName_A" and "rangeName_B" for categories "A"
and "B" and ranges that are set on RooRealVars.

This is the first implementation. Testing and documentation
changes will be added in the coming commits.