Finish rangeName in FitUtil::Expectation impl.
rangeName has been implemented into the Expectation method of
FitUtil. This allows for fitting simultaneously over different
ranges per pdf in RooSimultaneous.

Additionally a test has been written that checks if this
implementation works properly, and ALSO whether RooDataHist
integrates properly on the same range, compared to TH1.

A small note: the functionality of FitUtil::GetRange has been
slightly altered in the way it get the upper bin, but the tests have
not failed so far.