Change effmass and sumfile paths to new format in asimov dir
The paths of the Summary files and Effective Mass files have been
changed to the new format. This means that the basedir still uses
the 23m production from 2018-04, the sensitivity correction code
uses the 23m in the normal scripts, and the sensitivity correction
code uses 20m in the qsub code.