Added PyROOT support for MONA.
To this end, I added another function `EventFilter::AddCut`, which takes a string name of the SummaryEvent function and a string name for the comparison operator. This is required, as python does not work well with function pointers. The function names and operators are stored in private member maps of EventFilter. If SummaryEvent format is changed, the function name map should be updated to accommodate the changes.

I've added a demonstration macro to `macros/`.